backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
djoser==2.2.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
